requirements.txt
python-dotenv==1.0.0
streamlit==1.31.1
pandas==1.5.3
numpy==1.23.5
langchain==0.1.8
langchain-experimental==0.0.52
langchain-google-genai==0.0.9
langchain-google-vertexai==0.0.5
langchain-openai==0.0.6
Pillow==9.5.0
matplotlib==3.7.1
seaborn==0.12.2
PyQt5==5.15.10
google-search-results==2.4.2

